Fix security and performance issues from code audit
Security:
- Fix CSRF Origin validation (require header on exempt endpoints)
- Atomic password reset token consumption (prevent race condition)
- Viewer role checks on tracked-changes, comments, and GitHub push/pull
- WS message validation (comment text length, tracked-change status)
- Fetch timeouts on external APIs (CrossRef, Zotero)
- Downgrade Zotero key log from info to debug
- Fix decrypt() input validation
- Fix WS room leak on project re-join
- Add Redis error handlers, shutdown timeout

Performance:
- Add response compression middleware
- Fix SELECT * in permission-check queries
- Reduce redundant doc.toString() in editor
- Add pool error handler in db.js

React fixes:
- Fix setState during render in AuthPage
- Fix setState inside useMemo in BibEnrichModal
- Add null guard in Avatar component
- Fix Uint16Array overflow in lineDiff
- Clear all editor timeouts on unmount

Also: gitSync clearRemoteAuth in finally blocks, structured logging
for ZIP errors, ignore projects/ dir (compilation artifacts).

+++ client/src/utils/lineDiff.js
@@ -24,7 +24,7 @@ export default function lineDiff(oldText, newText) {
 
 function lcsLineDiff(oldLines, newLines, m, n) {
   const w = n + 1;
-  const dp = new Uint16Array((m + 1) * (n + 1));
+  const dp = new Uint32Array((m + 1) * (n + 1));
   for (let i = m - 1; i >= 0; i--) {
     for (let j = n - 1; j >= 0; j--) {
       if (oldLines[i] === newLines[j]) {
